+++ src/built_ins/close.rs
@@ -32,7 +32,7 @@ pub mod parser {
 }
 
 pub mod linter {
-    use crate::common::{QError, QErrorNode, ToErrorEnvelopeNoPos};
+    use crate::common::QErrorNode;
     use crate::linter::arg_validation::ArgValidation;
     use crate::parser::ExpressionNode;
 
@@ -45,10 +45,24 @@ pub mod linter {
 }
 
 pub mod interpreter {
-    use crate::common::QError;
+    use crate::common::{FileHandle, QError};
     use crate::interpreter::interpreter_trait::InterpreterTrait;
+    use crate::interpreter::utils::to_file_handle;
 
     pub fn run<S: InterpreterTrait>(interpreter: &mut S) -> Result<(), QError> {
+        let file_handles: Vec<FileHandle> = interpreter
+            .context()
+            .variables()
+            .iter()
+            .map(to_file_handle)
+            .collect::<Result<Vec<FileHandle>, QError>>()?;
+        if file_handles.is_empty() {
+            interpreter.file_manager().close_all();
+        } else {
+            for file_handle in file_handles {
+                interpreter.file_manager().close(&file_handle);
+            }
+        }
         Ok(())
     }
 }
@@ -58,6 +72,7 @@ mod tests {
     use crate::assert_parser_err;
     use crate::built_ins::BuiltInSub;
     use crate::common::*;
+    use crate::interpreter::test_utils::*;
     use crate::parser::test_utils::*;
     use crate::parser::*;
 
@@ -272,5 +287,37 @@ mod tests {
                     .at_rc(1, 10)
             ]
         );
+    }
+
+    #[test]
+    fn test_close_not_opened_file_is_allowed() {
+        interpret("CLOSE 1");
+    }
+
+    #[test]
+    fn test_close_allows_to_open_again() {
+        let input = r#"
+        OPEN "a.txt" FOR OUTPUT AS #1
+        CLOSE #1
+        OPEN "a.txt" FOR OUTPUT AS #1
+        CLOSE #1
+        "#;
+        interpret(input);
+        std::fs::remove_file("a.txt").unwrap_or(());
+    }
+
+    #[test]
+    fn test_close_without_args_closes_all_files() {
+        let input = r#"
+        OPEN "a.txt" FOR OUTPUT AS #1
+        OPEN "b.txt" FOR OUTPUT AS #2
+        CLOSE
+        OPEN "a.txt" FOR OUTPUT AS #1
+        OPEN "b.txt" FOR OUTPUT AS #2
+        CLOSE
+        "#;
+        interpret(input);
+        std::fs::remove_file("a.txt").unwrap_or(());
+        std::fs::remove_file("b.txt").unwrap_or(());
     }
 }
