+++ src/built_ins/environ_sub.rs
@@ -1,20 +1,5 @@
-pub mod parser {
-    use crate::built_ins::BuiltInSub;
-    use crate::common::*;
-    use crate::parser::pc::*;
-    use crate::parser::pc_specific::*;
-    use crate::parser::*;
-
-    pub fn parse<R>() -> impl Parser<R, Output = Statement>
-    where
-        R: Reader<Item = char, Err = QError> + HasLocation + 'static,
-    {
-        keyword_p(Keyword::Close).map(|_| Statement::BuiltInSubCall(BuiltInSub::Close, vec![]))
-    }
-}
-
 pub mod linter {
-    use crate::common::{QError, QErrorNode, ToErrorEnvelopeNoPos};
+    use crate::common::QErrorNode;
     use crate::linter::arg_validation::ArgValidation;
     use crate::parser::ExpressionNode;
 
@@ -26,11 +11,47 @@ pub mod linter {
 pub mod interpreter {
     use crate::common::QError;
     use crate::interpreter::interpreter_trait::InterpreterTrait;
+    use crate::interpreter::Stdlib;
 
     pub fn run<S: InterpreterTrait>(interpreter: &mut S) -> Result<(), QError> {
-        Ok(())
+        let s: &str = interpreter.context()[0].to_str_unchecked();
+        let parts: Vec<&str> = s.split("=").collect();
+        if parts.len() != 2 {
+            Err(QError::from("Invalid expression. Must be name=value."))
+        } else {
+            let name = parts[0].to_string();
+            let value = parts[1].to_string();
+            interpreter.stdlib_mut().set_env_var(name, value);
+            Ok(())
+        }
     }
 }
 
 #[cfg(test)]
-mod tests {}
+mod tests {
+    use crate::interpreter::interpreter_trait::InterpreterTrait;
+    use crate::interpreter::test_utils::interpret;
+    use crate::interpreter::Stdlib;
+
+    #[test]
+    fn test_sub_call_environ() {
+        let program = r#"
+        ENVIRON "FOO=BAR"
+        "#;
+        let interpreter = interpret(program);
+        assert_eq!(interpreter.stdlib().get_env_var(&"FOO".to_string()), "BAR");
+    }
+
+    #[test]
+    fn test_sub_call_environ_by_ref() {
+        let program = r#"
+        A$ = "FOO1=BAR2"
+        ENVIRON A$
+        "#;
+        let interpreter = interpret(program);
+        assert_eq!(
+            interpreter.stdlib().get_env_var(&"FOO1".to_string()),
+            "BAR2"
+        );
+    }
+}
