+++ src/interpreter/mod.rs
@@ -2,16 +2,14 @@ mod arguments;
 mod built_ins;
 pub mod context;
 mod default_stdlib;
-mod input;
 pub mod interpreter;
 pub mod interpreter_trait;
-mod io;
+pub mod io;
 mod lpt1_write;
 mod print;
-mod printer;
 mod read_input;
 mod registers;
-mod stdlib;
+pub mod utils;
 mod variables;
 mod write_printer;
 
@@ -19,3 +17,15 @@ mod write_printer;
 pub mod test_utils;
 #[cfg(test)]
 mod tests;
+
+/// The standard functions that QBasic offers
+pub trait Stdlib {
+    /// Implementation of SYSTEM
+    fn system(&self);
+
+    /// Gets an environment variable (used by built-in function ENVIRON$)
+    fn get_env_var(&self, name: &str) -> String;
+
+    /// Sets an environment variable (used by built-in sub ENVIRON)
+    fn set_env_var(&mut self, name: String, value: String);
+}
